refactor: split game map area turn feed into label and entry parts (#3878)
Extract turn-feed display labels and map-locate helpers into a dedicated
library part mixin; deduplicate LocateMapTileEvent tap handlers in the
entry builder for Phase 3 map_state modularization.
