+++ app/lib/features/game/flame/map_state/game_map_area.dart
@@ -66,6 +66,7 @@ part 'game_map_area_state_base.dart';
 part 'game_map_area_selection.dart';
 part 'game_map_area_view.dart';
 part 'game_map_area_turn_resolution.dart';
+part 'game_map_area_turn_feed_labels.dart';
 part 'game_map_area_turn_feed.dart';
 part 'game_map_area_events.dart';
 part 'game_map_area_e2e.dart';
@@ -96,6 +97,7 @@ class _GameMapAreaState extends ConsumerState<GameMapArea>
         _GameMapAreaSelection,
         _GameMapAreaView,
         _GameMapAreaTurnResolution,
+        _GameMapAreaTurnFeedLabels,
         _GameMapAreaTurnFeed,
         _GameMapAreaEvents,
         _GameMapAreaE2e,
