+++ app/lib/features/game/flame/map_state/game_map_area_turn_feed.dart
@@ -1,92 +1,12 @@
 part of 'game_map_area.dart';
 
 /// Player turn-event feed for [GameMapArea]: turning resolved `GameToUIEvent`s
-/// into tappable feed entries and resolving the faction/province/sea-zone
-/// display labels they render (Refs #3699 Theme 3).
-mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase {
-  String _factionLabel(String id) =>
-      widget.game.factionDisplayNameById(id) ?? id;
-
-  String _provinceLabel(String fullProvinceId) =>
-      widget.game.worldState.tryGetProvince(fullProvinceId)?.displayName ??
-      fullProvinceId;
-
-  String _seaZoneLabel(String seaZoneId) {
-    return widget.game.worldState.seaZoneDisplayNameById[seaZoneId] ??
-        seaZoneId;
-  }
-
-  String _diplomacyOutcomeLine({
-    required String actorId,
-    required String targetId,
-    required String changeType,
-  }) {
-    final actor = _factionLabel(actorId);
-    final target = _factionLabel(targetId);
-    final normalized = changeType.toLowerCase();
-    return switch (normalized) {
-      'declare_war' => '$actor declared war on $target!',
-      'peace' => '$actor and $target signed peace!',
-      'alliance' => '$actor and $target formed an alliance!',
-      'break_alliance' => '$actor and $target broke their alliance!',
-      _ => '$actor and $target diplomacy changed! ${changeType.toUpperCase()}!',
-    };
-  }
-
-  Set<String> _seaZoneRegionCandidates(String seaZoneId) {
-    final regionFromPrefix = prefixedIdRegionSegment(seaZoneId);
-    if (regionFromPrefix != null && regionFromPrefix.isNotEmpty) {
-      return {regionFromPrefix};
-    }
-    final localSeaZoneId = prefixedIdLocalSegment(seaZoneId);
-    final fromPorts = <String>{};
-    for (final key in widget.game.worldState.portsByProvinceSeaboard.keys) {
-      final firstPipe = key.indexOf('|');
-      if (firstPipe <= 0 || firstPipe + 1 >= key.length) {
-        continue;
-      }
-      final lastPipe = key.lastIndexOf('|');
-      final keyRegion = key.substring(0, firstPipe);
-      final keySeaZone = key.substring(lastPipe + 1);
-      if (keySeaZone == localSeaZoneId && keyRegion.isNotEmpty) {
-        fromPorts.add(keyRegion);
-      }
-    }
-    final mapData = ref.read(gameServiceProvider).getMapData(widget.game.id);
-    final fromTopology = <String>{};
-    if (mapData == null) {
-      return fromPorts;
-    }
-    for (final entry in mapData.topologyByRegion.entries) {
-      if (entry.value.nodes.any(
-        (node) =>
-            node.type == TopologyNodeType.seaZone && node.id == localSeaZoneId,
-      )) {
-        fromTopology.add(entry.key);
-      }
-    }
-    return {...fromPorts, ...fromTopology};
-  }
-
-  String? _tileKeyForSeaZoneEvent(String seaZoneId) {
-    final candidates = _seaZoneRegionCandidates(seaZoneId);
-    if (candidates.length != 1) {
-      return null;
-    }
-    final regionId = candidates.first;
-    final mapData = ref.read(gameServiceProvider).getMapData(widget.game.id);
-    return tileKeyForNavalFleetAtSea(
-      game: widget.game,
-      regionId: regionId,
-      seaZoneId: seaZoneId,
-      tileMap: mapData?.tileMapByRegion[regionId],
-      regionTopology: mapData?.topologyByRegion[regionId],
-    );
-  }
-
-  ct_models.Province? _provinceByPrefixedId(String prefixedProvinceId) =>
-      widget.game.worldState.tryGetProvince(prefixedProvinceId);
-
+/// into tappable feed entries (Refs #3699 Theme 3).
+mixin _GameMapAreaTurnFeed
+    on
+        ConsumerState<GameMapArea>,
+        _GameMapAreaStateBase,
+        _GameMapAreaTurnFeedLabels {
   List<PlayerTurnEventFeedEntry> _feedEntries() {
     return _resolvedPlayerTurnEvents
         .map((event) {
@@ -100,23 +20,7 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
               PlayerTurnEventFeedEntry(
                 text:
                     '${_provinceLabel(provinceId)} battle resolved! ${_factionLabel(winnerId)} defeated ${_factionLabel(winnerId == attackerId ? defenderId : attackerId)}!',
-                onTap: () {
-                  final province = _provinceByPrefixedId(provinceId);
-                  if (province == null) return;
-                  final tileKey = tileKeyForProvinceLocation(
-                    widget.game,
-                    province,
-                  );
-                  if (tileKey == null) return;
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: province.regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateProvinceById(provinceId),
               ),
             ct_models.AppProvinceCapturedEvent(
               :final provinceId,
@@ -125,23 +29,7 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
               PlayerTurnEventFeedEntry(
                 text:
                     '${_provinceLabel(provinceId)} captured! ${_factionLabel(newOwnerId)} now controls it!',
-                onTap: () {
-                  final province = _provinceByPrefixedId(provinceId);
-                  if (province == null) return;
-                  final tileKey = tileKeyForProvinceLocation(
-                    widget.game,
-                    province,
-                  );
-                  if (tileKey == null) return;
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: province.regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateProvinceById(provinceId),
               ),
             ct_models.AppNavalCombatResultEvent(
               :final seaZoneId,
@@ -150,24 +38,7 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
               PlayerTurnEventFeedEntry(
                 text:
                     '${_seaZoneLabel(seaZoneId)} naval battle resolved! Outcome: $outcomeName!',
-                onTap: () {
-                  final tileKey = _tileKeyForSeaZoneEvent(seaZoneId);
-                  if (tileKey == null) {
-                    return;
-                  }
-                  final regionId = ct_models.Unit.regionIdFromTileKey(tileKey);
-                  if (regionId == null) {
-                    return;
-                  }
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateSeaZoneTile(seaZoneId),
               ),
             ct_models.AppDiplomacyChangeEvent(
               :final actorId,
@@ -197,65 +68,17 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
               PlayerTurnEventFeedEntry(
                 text:
                     '${_provinceLabel(provinceId)} work completed! ${workTarget.toUpperCase()} finished!',
-                onTap: () {
-                  final regionId = ct_models.Unit.regionIdFromTileKey(
-                    targetTileKey,
-                  );
-                  if (regionId == null) {
-                    return;
-                  }
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: targetTileKey,
-                          regionId: regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateTileKey(targetTileKey),
               ),
             ct_models.AppPlayerProvinceDiscoveredEvent(:final provinceId) =>
               PlayerTurnEventFeedEntry(
                 text: '${_provinceLabel(provinceId)} discovered!',
-                onTap: () {
-                  final province = _provinceByPrefixedId(provinceId);
-                  if (province == null) return;
-                  final tileKey = tileKeyForProvinceLocation(
-                    widget.game,
-                    province,
-                  );
-                  if (tileKey == null) return;
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: province.regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateProvinceById(provinceId),
               ),
             ct_models.AppPlayerSeaZoneDiscoveredEvent(:final seaZoneId) =>
               PlayerTurnEventFeedEntry(
                 text: '${_seaZoneLabel(seaZoneId)} discovered!',
-                onTap: () {
-                  final tileKey = _tileKeyForSeaZoneEvent(seaZoneId);
-                  if (tileKey == null) {
-                    return;
-                  }
-                  final regionId = ct_models.Unit.regionIdFromTileKey(tileKey);
-                  if (regionId == null) {
-                    return;
-                  }
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateSeaZoneTile(seaZoneId),
               ),
             ct_models.AppOvertureAdvancedEvent(
               :final offererGpId,
@@ -275,23 +98,7 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
                 text: _mapPlayerId == territoryOwnerId
                     ? '${_provinceLabel(provinceId)} — enemy spy from ${_factionLabel(spyOwnerId)} caught and eliminated!'
                     : 'Spy caught in ${_provinceLabel(provinceId)}! ${_factionLabel(territoryOwnerId)} eliminated your agent!',
-                onTap: () {
-                  final province = _provinceByPrefixedId(provinceId);
-                  if (province == null) return;
-                  final tileKey = tileKeyForProvinceLocation(
-                    widget.game,
-                    province,
-                  );
-                  if (tileKey == null) return;
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: province.regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateProvinceById(provinceId),
               ),
             ct_models.AppSpyDefectedEvent(
               :final provinceId,
@@ -302,23 +109,7 @@ mixin _GameMapAreaTurnFeed on ConsumerState<GameMapArea>, _GameMapAreaStateBase 
                 text: _mapPlayerId == newOwnerId
                     ? '${_provinceLabel(provinceId)} — enemy spy from ${_factionLabel(previousOwnerId)} defected to your side!'
                     : 'Spy defected in ${_provinceLabel(provinceId)}! Agent joined ${_factionLabel(newOwnerId)}!',
-                onTap: () {
-                  final province = _provinceByPrefixedId(provinceId);
-                  if (province == null) return;
-                  final tileKey = tileKeyForProvinceLocation(
-                    widget.game,
-                    province,
-                  );
-                  if (tileKey == null) return;
-                  ref
-                      .read(appEventBusProvider)
-                      .emit(
-                        ct_models.LocateMapTileEvent(
-                          tileKey: tileKey,
-                          regionId: province.regionId,
-                        ),
-                      );
-                },
+                onTap: () => _locateProvinceById(provinceId),
               ),
             _ => const PlayerTurnEventFeedEntry(text: 'Event resolved!'),
           };
